server/requirements.txt
Flask
pika
redis
Flask-SQLAlchemy
flask-migrate
mysqlclient
sqlalchemy-utils
redisearch